+++ src/main/java/net/ilexiconn/showcase/server/message/MessageUpdateScale.java
@@ -7,7 +7,7 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.BlockPos;
 
 public class MessageUpdateScale extends AbstractMessage<MessageUpdateScale> {
-    public float modelScale;
+    public int modelScale;
     public BlockPos blockPos;
 
     public MessageUpdateScale() {
@@ -15,7 +15,7 @@ public class MessageUpdateScale extends AbstractMessage<MessageUpdateScale> {
     }
 
 
-    public MessageUpdateScale(float rotation, BlockPos pos) {
+    public MessageUpdateScale(int rotation, BlockPos pos) {
         modelScale = rotation;
         blockPos = pos;
     }
@@ -29,12 +29,12 @@ public class MessageUpdateScale extends AbstractMessage<MessageUpdateScale> {
     }
 
     public void fromBytes(ByteBuf buf) {
-        modelScale = buf.readFloat();
+        modelScale = buf.readInt();
         blockPos = new BlockPos(buf.readInt(), buf.readInt(), buf.readInt());
     }
 
     public void toBytes(ByteBuf buf) {
-        buf.writeFloat(modelScale);
+        buf.writeInt(modelScale);
         buf.writeInt(blockPos.getX());
         buf.writeInt(blockPos.getY());
         buf.writeInt(blockPos.getZ());
